Second Commit
I have added 5 more help files here. There are still some questions
I’ll have for Jeff about how what the data is supposed to represent for
some of them, but otherwise I feel confident that this is correct so
far and I look forward to looking it over Thursday with the rest of the
team. I will be committing the other help files tomorrow.